+++ rust/src/frb_generated.rs
@@ -38,7 +38,7 @@ flutter_rust_bridge::frb_generated_boilerplate!(
     default_rust_auto_opaque = RustAutoOpaqueMoi,
 );
 pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_VERSION: &str = "2.12.0";
-pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = 8087097;
+pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = -1294523770;
 
 // Section: executor
 
@@ -291,6 +291,36 @@ fn wire__crate__api__ffi_api__copy_range_as_markdown_impl(
         },
     )
 }
+fn wire__crate__api__ffi_api__copy_whole_note_as_markdown_impl(
+    ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
+    rust_vec_len_: i32,
+    data_len_: i32,
+) -> flutter_rust_bridge::for_generated::WireSyncRust2DartSse {
+    FLUTTER_RUST_BRIDGE_HANDLER.wrap_sync::<flutter_rust_bridge::for_generated::SseCodec, _>(
+        flutter_rust_bridge::for_generated::TaskInfo {
+            debug_name: "copy_whole_note_as_markdown",
+            port: None,
+            mode: flutter_rust_bridge::for_generated::FfiCallMode::Sync,
+        },
+        move || {
+            let message = unsafe {
+                flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(
+                    ptr_,
+                    rust_vec_len_,
+                    data_len_,
+                )
+            };
+            let mut deserializer =
+                flutter_rust_bridge::for_generated::SseDeserializer::new(message);
+            let api_note_id = <String>::sse_decode(&mut deserializer);
+            deserializer.end();
+            transform_result_sse::<_, crate::error::AppError>((move || {
+                let output_ok = crate::api::ffi_api::copy_whole_note_as_markdown(api_note_id)?;
+                Ok(output_ok)
+            })())
+        },
+    )
+}
 fn wire__crate__api__ffi_api__create_directory_impl(
     port_: flutter_rust_bridge::for_generated::MessagePort,
     ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
@@ -531,6 +561,36 @@ fn wire__crate__api__ffi_api__delete_range_impl(
             deserializer.end();
             transform_result_sse::<_, crate::error::AppError>((move || {
                 let output_ok = crate::api::ffi_api::delete_range(api_note_id, api_range)?;
+                Ok(output_ok)
+            })())
+        },
+    )
+}
+fn wire__crate__api__ffi_api__delete_whole_note_impl(
+    ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
+    rust_vec_len_: i32,
+    data_len_: i32,
+) -> flutter_rust_bridge::for_generated::WireSyncRust2DartSse {
+    FLUTTER_RUST_BRIDGE_HANDLER.wrap_sync::<flutter_rust_bridge::for_generated::SseCodec, _>(
+        flutter_rust_bridge::for_generated::TaskInfo {
+            debug_name: "delete_whole_note",
+            port: None,
+            mode: flutter_rust_bridge::for_generated::FfiCallMode::Sync,
+        },
+        move || {
+            let message = unsafe {
+                flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(
+                    ptr_,
+                    rust_vec_len_,
+                    data_len_,
+                )
+            };
+            let mut deserializer =
+                flutter_rust_bridge::for_generated::SseDeserializer::new(message);
+            let api_note_id = <String>::sse_decode(&mut deserializer);
+            deserializer.end();
+            transform_result_sse::<_, crate::error::AppError>((move || {
+                let output_ok = crate::api::ffi_api::delete_whole_note(api_note_id)?;
                 Ok(output_ok)
             })())
         },
@@ -1172,6 +1232,38 @@ fn wire__crate__api__ffi_api__replace_range_impl(
             transform_result_sse::<_, crate::error::AppError>((move || {
                 let output_ok =
                     crate::api::ffi_api::replace_range(api_note_id, api_range, api_replacement)?;
+                Ok(output_ok)
+            })())
+        },
+    )
+}
+fn wire__crate__api__ffi_api__replace_whole_note_impl(
+    ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
+    rust_vec_len_: i32,
+    data_len_: i32,
+) -> flutter_rust_bridge::for_generated::WireSyncRust2DartSse {
+    FLUTTER_RUST_BRIDGE_HANDLER.wrap_sync::<flutter_rust_bridge::for_generated::SseCodec, _>(
+        flutter_rust_bridge::for_generated::TaskInfo {
+            debug_name: "replace_whole_note",
+            port: None,
+            mode: flutter_rust_bridge::for_generated::FfiCallMode::Sync,
+        },
+        move || {
+            let message = unsafe {
+                flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(
+                    ptr_,
+                    rust_vec_len_,
+                    data_len_,
+                )
+            };
+            let mut deserializer =
+                flutter_rust_bridge::for_generated::SseDeserializer::new(message);
+            let api_note_id = <String>::sse_decode(&mut deserializer);
+            let api_replacement = <String>::sse_decode(&mut deserializer);
+            deserializer.end();
+            transform_result_sse::<_, crate::error::AppError>((move || {
+                let output_ok =
+                    crate::api::ffi_api::replace_whole_note(api_note_id, api_replacement)?;
                 Ok(output_ok)
             })())
         },
@@ -2193,36 +2285,36 @@ fn pde_ffi_dispatcher_primary_impl(
         1 => wire__crate__api__auth__authenticate_workspace_impl(port, ptr, rust_vec_len, data_len),
         2 => wire__crate__api__ffi_api__backlinks_impl(port, ptr, rust_vec_len, data_len),
         4 => wire__crate__api__ffi_api__close_note_impl(port, ptr, rust_vec_len, data_len),
-        8 => wire__crate__api__ffi_api__create_directory_impl(port, ptr, rust_vec_len, data_len),
-        9 => wire__crate__api__ffi_api__create_link_target_impl(port, ptr, rust_vec_len, data_len),
-        10 => wire__crate__api__ffi_api__create_note_impl(port, ptr, rust_vec_len, data_len),
-        12 => wire__crate__api__ffi_api__delete_directory_impl(port, ptr, rust_vec_len, data_len),
-        13 => wire__crate__api__ffi_api__delete_note_impl(port, ptr, rust_vec_len, data_len),
-        15 => {
+        9 => wire__crate__api__ffi_api__create_directory_impl(port, ptr, rust_vec_len, data_len),
+        10 => wire__crate__api__ffi_api__create_link_target_impl(port, ptr, rust_vec_len, data_len),
+        11 => wire__crate__api__ffi_api__create_note_impl(port, ptr, rust_vec_len, data_len),
+        13 => wire__crate__api__ffi_api__delete_directory_impl(port, ptr, rust_vec_len, data_len),
+        14 => wire__crate__api__ffi_api__delete_note_impl(port, ptr, rust_vec_len, data_len),
+        17 => {
             wire__crate__api__ffi_api__find_notes_by_title_impl(port, ptr, rust_vec_len, data_len)
         }
-        16 => wire__crate__api__ffi_api__flush_note_impl(port, ptr, rust_vec_len, data_len),
-        18 => wire__crate__api__simple__init_app_impl(port, ptr, rust_vec_len, data_len),
-        20 => wire__crate__api__ffi_api__link_completions_impl(port, ptr, rust_vec_len, data_len),
-        22 => wire__crate__api__ffi_api__move_note_impl(port, ptr, rust_vec_len, data_len),
-        24 => wire__crate__api__ffi_api__open_note_impl(port, ptr, rust_vec_len, data_len),
-        25 => wire__crate__api__ffi_api__open_or_create_local_workspace_impl(
+        18 => wire__crate__api__ffi_api__flush_note_impl(port, ptr, rust_vec_len, data_len),
+        20 => wire__crate__api__simple__init_app_impl(port, ptr, rust_vec_len, data_len),
+        22 => wire__crate__api__ffi_api__link_completions_impl(port, ptr, rust_vec_len, data_len),
+        24 => wire__crate__api__ffi_api__move_note_impl(port, ptr, rust_vec_len, data_len),
+        26 => wire__crate__api__ffi_api__open_note_impl(port, ptr, rust_vec_len, data_len),
+        27 => wire__crate__api__ffi_api__open_or_create_local_workspace_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        26 => wire__crate__api__ffi_api__open_workspace_impl(port, ptr, rust_vec_len, data_len),
-        27 => wire__crate__api__ffi_api__pending_drafts_impl(port, ptr, rust_vec_len, data_len),
-        28 => wire__crate__api__ffi_api__reindex_workspace_impl(port, ptr, rust_vec_len, data_len),
-        29 => wire__crate__api__ffi_api__reload_note_impl(port, ptr, rust_vec_len, data_len),
-        30 => wire__crate__api__ffi_api__rename_directory_impl(port, ptr, rust_vec_len, data_len),
-        31 => wire__crate__api__ffi_api__rename_note_impl(port, ptr, rust_vec_len, data_len),
-        34 => {
+        28 => wire__crate__api__ffi_api__open_workspace_impl(port, ptr, rust_vec_len, data_len),
+        29 => wire__crate__api__ffi_api__pending_drafts_impl(port, ptr, rust_vec_len, data_len),
+        30 => wire__crate__api__ffi_api__reindex_workspace_impl(port, ptr, rust_vec_len, data_len),
+        31 => wire__crate__api__ffi_api__reload_note_impl(port, ptr, rust_vec_len, data_len),
+        32 => wire__crate__api__ffi_api__rename_directory_impl(port, ptr, rust_vec_len, data_len),
+        33 => wire__crate__api__ffi_api__rename_note_impl(port, ptr, rust_vec_len, data_len),
+        37 => {
             wire__crate__api__ffi_api__resolve_link_target_impl(port, ptr, rust_vec_len, data_len)
         }
-        35 => wire__crate__api__ffi_api__search_notes_impl(port, ptr, rust_vec_len, data_len),
-        38 => wire__crate__api__ffi_api__workspace_tree_impl(port, ptr, rust_vec_len, data_len),
+        38 => wire__crate__api__ffi_api__search_notes_impl(port, ptr, rust_vec_len, data_len),
+        41 => wire__crate__api__ffi_api__workspace_tree_impl(port, ptr, rust_vec_len, data_len),
         _ => unreachable!(),
     }
 }
@@ -2239,18 +2331,23 @@ fn pde_ffi_dispatcher_sync_impl(
         5 => wire__crate__api__ffi_api__commit_block_impl(ptr, rust_vec_len, data_len),
         6 => wire__crate__api__ffi_api__continue_block_after_impl(ptr, rust_vec_len, data_len),
         7 => wire__crate__api__ffi_api__copy_range_as_markdown_impl(ptr, rust_vec_len, data_len),
-        11 => wire__crate__api__ffi_api__delete_block_impl(ptr, rust_vec_len, data_len),
-        14 => wire__crate__api__ffi_api__delete_range_impl(ptr, rust_vec_len, data_len),
-        17 => wire__crate__api__ffi_api__get_block_source_impl(ptr, rust_vec_len, data_len),
-        19 => wire__crate__api__ffi_api__insert_block_impl(ptr, rust_vec_len, data_len),
-        21 => {
+        8 => {
+            wire__crate__api__ffi_api__copy_whole_note_as_markdown_impl(ptr, rust_vec_len, data_len)
+        }
+        12 => wire__crate__api__ffi_api__delete_block_impl(ptr, rust_vec_len, data_len),
+        15 => wire__crate__api__ffi_api__delete_range_impl(ptr, rust_vec_len, data_len),
+        16 => wire__crate__api__ffi_api__delete_whole_note_impl(ptr, rust_vec_len, data_len),
+        19 => wire__crate__api__ffi_api__get_block_source_impl(ptr, rust_vec_len, data_len),
+        21 => wire__crate__api__ffi_api__insert_block_impl(ptr, rust_vec_len, data_len),
+        23 => {
             wire__crate__api__ffi_api__merge_block_with_previous_impl(ptr, rust_vec_len, data_len)
         }
-        23 => wire__crate__api__ffi_api__note_write_status_impl(ptr, rust_vec_len, data_len),
-        32 => wire__crate__api__ffi_api__replace_range_impl(ptr, rust_vec_len, data_len),
-        33 => wire__crate__api__ffi_api__resolve_block_caret_impl(ptr, rust_vec_len, data_len),
-        36 => wire__crate__api__ffi_api__split_block_impl(ptr, rust_vec_len, data_len),
-        37 => wire__crate__api__ffi_api__update_block_impl(ptr, rust_vec_len, data_len),
+        25 => wire__crate__api__ffi_api__note_write_status_impl(ptr, rust_vec_len, data_len),
+        34 => wire__crate__api__ffi_api__replace_range_impl(ptr, rust_vec_len, data_len),
+        35 => wire__crate__api__ffi_api__replace_whole_note_impl(ptr, rust_vec_len, data_len),
+        36 => wire__crate__api__ffi_api__resolve_block_caret_impl(ptr, rust_vec_len, data_len),
+        39 => wire__crate__api__ffi_api__split_block_impl(ptr, rust_vec_len, data_len),
+        40 => wire__crate__api__ffi_api__update_block_impl(ptr, rust_vec_len, data_len),
         _ => unreachable!(),
     }
 }